Moved clipboard code out
